backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==22.0.0

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.13.3

# Auth (Google OAuth + JWT sessions)
passlib[bcrypt]
python-jose[cryptography]==3.3.0
authlib==1.3.2

# AI
google-generativeai==0.8.3

# Validation & Config
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# HTTP
httpx==0.27.2

# CORS
python-multipart==0.0.12

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
aiosqlite==0.20.0
